+++ CHANGES.rst
@@ -248,6 +248,10 @@ API Changes
     ``DAOStarFinder`` and ``IRAFStarFinder`` are now deprecated. Use the
     ``sharpness_range=(lower, upper)`` tuple keyword instead. [#2216]
 
+  - The ``roundlo`` and ``roundhi`` keyword arguments for
+    ``DAOStarFinder`` and ``IRAFStarFinder`` are now deprecated. Use the
+    ``roundness_range=(lower, upper)`` tuple keyword instead. [#2216]
+
 - ``photutils.profiles``
 
   - Cached Gaussian fits to the radial profile are now automatically